update nginx/dep updates/fix eslint/change line endings

+++ backend/internal/access-list.js
@@ -1,67 +1,65 @@
-const _                     = require('lodash');
-const fs                    = require('fs');
-const batchflow             = require('batchflow');
-const logger                = require('../logger').access;
-const error                 = require('../lib/error');
-const utils                 = require('../lib/utils');
-const accessListModel       = require('../models/access_list');
-const accessListAuthModel   = require('../models/access_list_auth');
+const _ = require('lodash');
+const fs = require('fs');
+const batchflow = require('batchflow');
+const logger = require('../logger').access;
+const error = require('../lib/error');
+const utils = require('../lib/utils');
+const accessListModel = require('../models/access_list');
+const accessListAuthModel = require('../models/access_list_auth');
 const accessListClientModel = require('../models/access_list_client');
-const proxyHostModel        = require('../models/proxy_host');
-const internalAuditLog      = require('./audit-log');
-const internalNginx         = require('./nginx');
+const proxyHostModel = require('../models/proxy_host');
+const internalAuditLog = require('./audit-log');
+const internalNginx = require('./nginx');
 
-function omissions () {
+function omissions() {
 	return ['is_deleted'];
 }
 
 const internalAccessList = {
-
 	/**
 	 * @param   {Access}  access
 	 * @param   {Object}  data
 	 * @returns {Promise}
 	 */
 	create: (access, data) => {
-		return access.can('access_lists:create', data)
-			.then((/*access_data*/) => {
+		return access
+			.can('access_lists:create', data)
+			.then((/* access_data */) => {
 				return accessListModel
 					.query()
 					.insertAndFetch({
-						name:          data.name,
-						satisfy_any:   data.satisfy_any,
-						pass_auth:     data.pass_auth,
-						owner_user_id: access.token.getUserId(1)
+						name: data.name,
+						satisfy_any: data.satisfy_any,
+						pass_auth: data.pass_auth,
+						owner_user_id: access.token.getUserId(1),
 					})
 					.then(utils.omitRow(omissions()));
 			})
 			.then((row) => {
 				data.id = row.id;
 
-				let promises = [];
+				const promises = [];
 
 				// Now add the items
 				data.items.map((item) => {
-					promises.push(accessListAuthModel
-						.query()
-						.insert({
+					promises.push(
+						accessListAuthModel.query().insert({
 							access_list_id: row.id,
-							username:       item.username,
-							password:       item.password
-						})
+							username: item.username,
+							password: item.password,
+						}),
 					);
 				});
 
 				// Now add the clients
 				if (typeof data.clients !== 'undefined' && data.clients) {
 					data.clients.map((client) => {
-						promises.push(accessListClientModel
-							.query()
-							.insert({
+						promises.push(
+							accessListClientModel.query().insert({
 								access_list_id: row.id,
-								address:        client.address,
-								directive:      client.directive
-							})
+								address: client.address,
+								directive: client.directive,
+							}),
 						);
 					});
 				}
@@ -70,16 +68,21 @@ const internalAccessList = {
 			})
 			.then(() => {
 				// re-fetch with expansions
-				return internalAccessList.get(access, {
-					id:     data.id,
-					expand: ['owner', 'items', 'clients', 'proxy_hosts.access_list.[clients,items]']
-				}, true /* <- skip masking */);
+				return internalAccessList.get(
+					access,
+					{
+						id: data.id,
+						expand: ['owner', 'items', 'clients', 'proxy_hosts.access_list.[clients,items]'],
+					},
+					true /* <- skip masking */,
+				);
 			})
 			.then((row) => {
 				// Audit log
 				data.meta = _.assign({}, data.meta || {}, row.meta);
 
-				return internalAccessList.build(row)
+				return internalAccessList
+					.build(row)
 					.then(() => {
 						if (row.proxy_host_count) {
 							return internalNginx.bulkGenerateConfigs('proxy_host', row.proxy_hosts);
@@ -88,10 +91,10 @@ const internalAccessList = {
 					.then(() => {
 						// Add to audit log
 						return internalAuditLog.add(access, {
-							action:      'created',
+							action: 'created',
 							object_type: 'access-list',
-							object_id:   row.id,
-							meta:        internalAccessList.maskItems(data)
+							object_id: row.id,
+							meta: internalAccessList.maskItems(data),
 						});
 					})
 					.then(() => {
@@ -109,9 +112,10 @@ const internalAccessList = {
 	 * @return {Promise}
 	 */
 	update: (access, data) => {
-		return access.can('access_lists:update', data.id)
-			.then((/*access_data*/) => {
-				return internalAccessList.get(access, {id: data.id});
+		return access
+			.can('access_lists:update', data.id)
+			.then((/* access_data */) => {
+				return internalAccessList.get(access, { id: data.id });
 			})
 			.then((row) => {
 				if (row.id !== data.id) {
@@ -122,31 +126,27 @@ const internalAccessList = {
 			.then(() => {
 				// patch name if specified
 				if (typeof data.name !== 'undefined' && data.name) {
-					return accessListModel
-						.query()
-						.where({id: data.id})
-						.patch({
-							name:        data.name,
-							satisfy_any: data.satisfy_any,
-							pass_auth:   data.pass_auth,
-						});
+					return accessListModel.query().where({ id: data.id }).patch({
+						name: data.name,
+						satisfy_any: data.satisfy_any,
+						pass_auth: data.pass_auth,
+					});
 				}
 			})
 			.then(() => {
 				// Check for items and add/update/remove them
 				if (typeof data.items !== 'undefined' && data.items) {
-					let promises      = [];
-					let items_to_keep = [];
+					const promises = [];
+					const items_to_keep = [];
 
 					data.items.map(function (item) {
 						if (item.password) {
-							promises.push(accessListAuthModel
-								.query()
-								.insert({
+							promises.push(
+								accessListAuthModel.query().insert({
 									access_list_id: data.id,
-									username:       item.username,
-									password:       item.password
-								})
+									username: item.username,
+									password: item.password,
+								}),
 							);
 						} else {
 							// This was supplied with an empty password, which means keep it but don't change the password
@@ -154,79 +154,76 @@ const internalAccessList = {
 						}
 					});
 
-					let query = accessListAuthModel
-						.query()
-						.delete()
-						.where('access_list_id', data.id);
+					const query = accessListAuthModel.query().delete().where('access_list_id', data.id);
 
 					if (items_to_keep.length) {
 						query.andWhere('username', 'NOT IN', items_to_keep);
 					}
 
-					return query
-						.then(() => {
-							// Add new items
-							if (promises.length) {
-								return Promise.all(promises);
-							}
-						});
+					return query.then(() => {
+						// Add new items
+						if (promises.length) {
+							return Promise.all(promises);
+						}
+					});
 				}
 			})
 			.then(() => {
 				// Check for clients and add/update/remove them
 				if (typeof data.clients !== 'undefined' && data.clients) {
-					let promises = [];
+					const promises = [];
 
 					data.clients.map(function (client) {
 						if (client.address) {
-							promises.push(accessListClientModel
-								.query()
-								.insert({
+							promises.push(
+								accessListClientModel.query().insert({
 									access_list_id: data.id,
-									address:        client.address,
-									directive:      client.directive
-								})
+									address: client.address,
+									directive: client.directive,
+								}),
 							);
 						}
 					});
 
-					let query = accessListClientModel
-						.query()
-						.delete()
-						.where('access_list_id', data.id);
+					const query = accessListClientModel.query().delete().where('access_list_id', data.id);
 
-					return query
-						.then(() => {
-							// Add new items
-							if (promises.length) {
-								return Promise.all(promises);
-							}
-						});
+					return query.then(() => {
+						// Add new items
+						if (promises.length) {
+							return Promise.all(promises);
+						}
+					});
 				}
 			})
 			.then(() => {
 				// Add to audit log
 				return internalAuditLog.add(access, {
-					action:      'updated',
+					action: 'updated',
 					object_type: 'access-list',
-					object_id:   data.id,
-					meta:        internalAccessList.maskItems(data)
+					object_id: data.id,
+					meta: internalAccessList.maskItems(data),
 				});
 			})
 			.then(() => {
 				// re-fetch with expansions
-				return internalAccessList.get(access, {
-					id:     data.id,
-					expand: ['owner', 'items', 'clients', 'proxy_hosts.[certificate,access_list.[clients,items]]']
-				}, true /* <- skip masking */);
+				return internalAccessList.get(
+					access,
+					{
+						id: data.id,
+						expand: ['owner', 'items', 'clients', 'proxy_hosts.[certificate,access_list.[clients,items]]'],
+					},
+					true /* <- skip masking */,
+				);
 			})
 			.then((row) => {
-				return internalAccessList.build(row)
+				return internalAccessList
+					.build(row)
 					.then(() => {
 						if (row.proxy_host_count) {
 							return internalNginx.bulkGenerateConfigs('proxy_host', row.proxy_hosts);
 						}
-					}).then(internalNginx.reload)
+					})
+					.then(internalNginx.reload)
 					.then(() => {
 						return internalAccessList.maskItems(row);
 					});
@@ -247,16 +244,10 @@ const internalAccessList = {
 			data = {};
 		}
 
-		return access.can('access_lists:get', data.id)
+		return access
+			.can('access_lists:get', data.id)
 			.then((access_data) => {
-				let query = accessListModel
-					.query()
-					.select('access_list.*', accessListModel.raw('COUNT(proxy_host.id) as proxy_host_count'))
-					.joinRaw('LEFT JOIN `proxy_host` ON `proxy_host`.`access_list_id` = `access_list`.`id` AND `proxy_host`.`is_deleted` = 0')
-					.where('access_list.is_deleted', 0)
-					.andWhere('access_list.id', data.id)
-					.allowGraph('[owner,items,clients,proxy_hosts.[certificate,access_list.[clients,items]]]')
-					.first();
+				const query = accessListModel.query().select('access_list.*', accessListModel.raw('COUNT(proxy_host.id) as proxy_host_count')).joinRaw('LEFT JOIN `proxy_host` ON `proxy_host`.`access_list_id` = `access_list`.`id` AND `proxy_host`.`is_deleted` = 0').where('access_list.is_deleted', 0).andWhere('access_list.id', data.id).allowGraph('[owner,items,clients,proxy_hosts.[certificate,access_list.[clients,items]]]').first();
 
 				if (access_data.permission_visibility !== 'all') {
 					query.andWhere('access_list.owner_user_id', access.token.getUserId(1));
@@ -291,9 +282,10 @@ const internalAccessList = {
 	 * @returns {Promise}
 	 */
 	delete: (access, data) => {
-		return access.can('access_lists:delete', data.id)
+		return access
+			.can('access_lists:delete', data.id)
 			.then(() => {
-				return internalAccessList.get(access, {id: data.id, expand: ['proxy_hosts', 'items', 'clients']});
+				return internalAccessList.get(access, { id: data.id, expand: ['proxy_hosts', 'items', 'clients'] });
 			})
 			.then((row) => {
 				if (!row) {
@@ -310,7 +302,7 @@ const internalAccessList = {
 					.query()
 					.where('id', row.id)
 					.patch({
-						is_deleted: 1
+						is_deleted: 1,
 					})
 					.then(() => {
 						// 2. update any proxy hosts that were using it (ignoring permissions)
@@ -318,7 +310,7 @@ const internalAccessList = {
 							return proxyHostModel
 								.query()
 								.where('access_list_id', '=', row.id)
-								.patch({access_list_id: 0})
+								.patch({ access_list_id: 0 })
 								.then(() => {
 									// 3. reconfigure those hosts, then reload nginx
 
@@ -336,21 +328,21 @@ const internalAccessList = {
 					})
 					.then(() => {
 						// delete the htpasswd file
-						let htpasswd_file = internalAccessList.getFilename(row);
+						const htpasswd_file = internalAccessList.getFilename(row);
 
 						try {
 							fs.unlinkSync(htpasswd_file);
-						} catch (err) {
+						} catch {
 							// do nothing
 						}
 					})
 					.then(() => {
 						// 4. audit log
 						return internalAuditLog.add(access, {
-							action:      'deleted',
+							action: 'deleted',
 							object_type: 'access-list',
-							object_id:   row.id,
-							meta:        _.omit(internalAccessList.maskItems(row), ['is_deleted', 'proxy_hosts'])
+							object_id: row.id,
+							meta: _.omit(internalAccessList.maskItems(row), ['is_deleted', 'proxy_hosts']),
 						});
 					});
 			})
@@ -368,16 +360,10 @@ const internalAccessList = {
 	 * @returns {Promise}
 	 */
 	getAll: (access, expand, search_query) => {
-		return access.can('access_lists:list')
+		return access
+			.can('access_lists:list')
 			.then((access_data) => {
-				let query = accessListModel
-					.query()
-					.select('access_list.*', accessListModel.raw('COUNT(proxy_host.id) as proxy_host_count'))
-					.joinRaw('LEFT JOIN `proxy_host` ON `proxy_host`.`access_list_id` = `access_list`.`id` AND `proxy_host`.`is_deleted` = 0')
-					.where('access_list.is_deleted', 0)
-					.groupBy('access_list.id')
-					.allowGraph('[owner,items,clients]')
-					.orderBy('access_list.name', 'ASC');
+				const query = accessListModel.query().select('access_list.*', accessListModel.raw('COUNT(proxy_host.id) as proxy_host_count')).joinRaw('LEFT JOIN `proxy_host` ON `proxy_host`.`access_list_id` = `access_list`.`id` AND `proxy_host`.`is_deleted` = 0').where('access_list.is_deleted', 0).groupBy('access_list.id').allowGraph('[owner,items,clients]').orderBy('access_list.name', 'ASC');
 
 				if (access_data.permission_visibility !== 'all') {
 					query.andWhere('access_list.owner_user_id', access.token.getUserId(1));
@@ -417,19 +403,15 @@ const internalAccessList = {
 	 * @returns {Promise}
 	 */
 	getCount: (user_id, visibility) => {
-		let query = accessListModel
-			.query()
-			.count('id as count')
-			.where('is_deleted', 0);
+		const query = accessListModel.query().count('id as count').where('is_deleted', 0);
 
 		if (visibility !== 'all') {
 			query.andWhere('owner_user_id', user_id);
 		}
 
-		return query.first()
-			.then((row) => {
-				return parseInt(row.count, 10);
-			});
+		return query.first().then((row) => {
+			return parseInt(row.count, 10);
+		});
 	},
 
 	/**
@@ -447,7 +429,7 @@ const internalAccessList = {
 					first_char = val.password.charAt(0);
 				}
 
-				list.items[idx].hint     = first_char + ('*').repeat(repeat_for);
+				list.items[idx].hint = first_char + '*'.repeat(repeat_for);
 				list.items[idx].password = '';
 			});
 		}
@@ -475,54 +457,55 @@ const internalAccessList = {
 		logger.info('Building Access file #' + list.id + ' for: ' + list.name);
 
 		return new Promise((resolve, reject) => {
-			let htpasswd_file = internalAccessList.getFilename(list);
+			const htpasswd_file = internalAccessList.getFilename(list);
 
 			// 1. remove any existing access file
 			try {
 				fs.unlinkSync(htpasswd_file);
-			} catch (err) {
+			} catch {
 				// do nothing
 			}
 
 			// 2. create empty access file
 			try {
-				fs.writeFileSync(htpasswd_file, '', {encoding: 'utf8'});
+				fs.writeFileSync(htpasswd_file, '', { encoding: 'utf8' });
 				resolve(htpasswd_file);
 			} catch (err) {
 				reject(err);
 			}
-		})
-			.then((htpasswd_file) => {
-				// 3. generate password for each user
-				if (list.items.length) {
-					return new Promise((resolve, reject) => {
-						batchflow(list.items).sequential()
-							.each((i, item, next) => {
-								if (typeof item.password !== 'undefined' && item.password.length) {
-									logger.info('Adding: ' + item.username);
+		}).then((htpasswd_file) => {
+			// 3. generate password for each user
+			if (list.items.length) {
+				return new Promise((resolve, reject) => {
+					batchflow(list.items)
+						.sequential()
+						.each((i, item, next) => {
+							if (typeof item.password !== 'undefined' && item.password.length) {
+								logger.info('Adding: ' + item.username);
 
-									utils.execFile('htpasswd', ['-b', htpasswd_file, item.username, item.password])
-										.then((/*result*/) => {
-											next();
-										})
-										.catch((err) => {
-											logger.error(err);
-											next(err);
-										});
-								}
-							})
-							.error((err) => {
-								logger.error(err);
-								reject(err);
-							})
-							.end((results) => {
-								logger.success('Built Access file #' + list.id + ' for: ' + list.name);
-								resolve(results);
-							});
-					});
-				}
-			});
-	}
+								utils
+									.execFile('htpasswd', ['-b', htpasswd_file, item.username, item.password])
+									.then((/* result */) => {
+										next();
+									})
+									.catch((err) => {
+										logger.error(err);
+										next(err);
+									});
+							}
+						})
+						.error((err) => {
+							logger.error(err);
+							reject(err);
+						})
+						.end((results) => {
+							logger.success('Built Access file #' + list.id + ' for: ' + list.name);
+							resolve(results);
+						});
+				});
+			}
+		});
+	},
 };
 
 module.exports = internalAccessList;
